requirements.txt
streamlit
audio-recorder-streamlit
openai
python-dotenv
gspread
oauth2client
pandas